Revert app content to v1.0.0 release state
Restores all tracked files to match the v1.0.0 tag (e37e1e3), undoing changes from the 24 commits merged since then.
