--- .github/workflows/flutter-android-release.yml
@@ -1,100 +1,0 @@
-name: Flutter Android Release
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Set up Java
-        uses: actions/setup-java@v4
-        with:
-          distribution: 'zulu'
-          java-version: '17'
-
-      - name: Set up Flutter
-        uses: subosito/flutter-action@v2
-        with:
-          channel: 'stable'
-
-      - name: Get version and build number
-        id: version
-        run: |
-          VERSION_NAME=$(grep 'version:' pubspec.yaml | cut -d ' ' -f 2 | cut -d '+' -f 1)
-          # استفاده از تعداد کامیت‌ها به عنوان شماره بیلد
-          BUILD_NUMBER=$(git rev-list --count HEAD)
-          
-          echo "VERSION_NAME=$VERSION_NAME" >> $GITHUB_ENV
-          echo "BUILD_NUMBER=$BUILD_NUMBER" >> $GITHUB_ENV
-          
-          # تغییر مهم: اضافه کردن شماره بیلد به نام تگ برای یکتا شدن
-          # مثال: v1.0.0-11
-          echo "VERSION=v$VERSION_NAME-$BUILD_NUMBER" >> $GITHUB_ENV
-
-      - name: Update pubspec.yaml with build number
-        run: |
-          sed -i "s/version: .*/version: ${{ env.VERSION_NAME }}+${{ env.BUILD_NUMBER }}/" pubspec.yaml
-
-      # آن بخش چک کردن تگ تکراری حذف شد چون با روش بالا تگ همیشه یکتا است
-      
-      - name: Install dependencies
-        run: flutter pub get
-
-      - name: Build APK
-        run: flutter build apk --release
-
-      - name: Rename APK
-        id: rename_apk
-        run: |
-          # نام فایل خروجی شامل ورژن و شماره بیلد خواهد بود
-          mv build/app/outputs/flutter-apk/app-release.apk app-${{ env.VERSION }}.apk
-          echo "APK_PATH=app-${{ env.VERSION }}.apk" >> $GITHUB_ENV
-
-      - name: Generate changelog
-        id: changelog
-        run: |
-          if git describe --tags --abbrev=0 >/dev/null 2>&1; then
-            CHANGELOG=$(git log $(git describe --tags --abbrev=0)..HEAD --pretty=format:"* %s")
-          else
-            CHANGELOG=$(git log --pretty=format:"* %s")
-          fi
-          # هندل کردن حالت خالی بودن لاگ برای جلوگیری از خطا
-          if [ -z "$CHANGELOG" ]; then
-            CHANGELOG="No major changes."
-          fi
-          
-          echo "CHANGELOG<<EOF" >> $GITHUB_ENV
-          echo "$CHANGELOG" >> $GITHUB_ENV
-          echo "EOF" >> $GITHUB_ENV
-
-      - name: Create Git tag
-        run: |
-          git config --global user.name 'github-actions[bot]'
-          git config --global user.email 'github-actions[bot]@users.noreply.github.com'
-          git tag -a "${{ env.VERSION }}" -m "Release ${{ env.VERSION }}"
-          git push origin "${{ env.VERSION }}"
-
-      - name: Create GitHub Release
-        uses: softprops/action-gh-release@v2
-        with:
-          files: ${{ env.APK_PATH }}
-          tag_name: ${{ env.VERSION }}
-          name: "Release ${{ env.VERSION }}"
-          body: |
-            Build Number: ${{ env.BUILD_NUMBER }}
-            Version: ${{ env.VERSION_NAME }}
-            
-            Changes:
-            ${{ env.CHANGELOG }}
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
