--- lib/core/constants.dart
@@ -1,1 +1,0 @@
-const String kTasksBox = 'ive_tasks_final_v6';
