--- lib/features/task_management/domain/entities/sub_task.dart
@@ -1,28 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-class SubTask extends Equatable {
-  final String id;
-  final String title;
-  final bool isCompleted;
-
-  const SubTask({
-    required this.id,
-    required this.title,
-    this.isCompleted = false,
-  });
-
-  SubTask copyWith({
-    String? id,
-    String? title,
-    bool? isCompleted,
-  }) {
-    return SubTask(
-      id: id ?? this.id,
-      title: title ?? this.title,
-      isCompleted: isCompleted ?? this.isCompleted,
-    );
-  }
-
-  @override
-  List<Object?> get props => [id, title, isCompleted];
-}
