--- lib/features/task_management/domain/repositories/task_repository.dart
@@ -1,6 +1,0 @@
-import 'package:app/features/task_management/domain/entities/task.dart';
-
-abstract class TaskRepository {
-  Future<List<Task>> getTasks();
-  Future<void> saveTasks(List<Task> tasks);
-}
