+++ pubspec.yaml
@@ -11,9 +11,7 @@ dependencies:
     sdk: flutter
   
   # State management
-  flutter_bloc: ^8.1.3
-  get_it: ^7.6.4
-  equatable: ^2.0.5
+  provider: ^6.1.2
   
   # Storage
   shared_preferences: ^2.2.2
@@ -24,8 +22,6 @@ dependencies:
   flutter_animate: ^4.5.0
   percent_indicator: ^4.2.3
   modal_bottom_sheet: ^3.0.0
-  fl_chart: ^0.68.0
-  collection: ^1.18.0
   
   # Utilities
   shamsi_date: ^1.0.0
@@ -34,15 +30,11 @@ dependencies:
   timezone: ^0.10.1
   flutter_local_notifications: ^18.0.1
   vibration: ^3.1.4
-  permission_handler: ^11.3.1
-  device_info_plus: ^10.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^4.0.0
-  bloc_test: ^9.1.5
-  mockito: ^5.4.4
   # برای ساخت آیکون اپلیکیشن از روی logo.png
   flutter_launcher_icons: ^0.13.1
